Refactor, make it cleaner
